# backend/src/api/swagger/swagger.yaml
...
openapi: 3.0.0
info:
  description: "OpenAPI specification for MemeGeneratorAPI"
  version: "1.0.0"
  title: "Meme Generator API"
paths:
...
  /memes/:
    get:
      summary: "Get public memes"
      tags:
        - Memes
      operationId: "getPublicMemes"
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Max number of items
        - in: query
          name: likes
          schema:
            type: integer
          description: Min number of likes meme needs to have
        - in: query
          name: views
          schema:
            type: integer
          description: Min number of views meme needs to have
        - in: query
          name: createdAfter
          schema:
            type: integer
          description: Meme needs to created before this date
        - in: query
          name: createdBefore
          schema:
            type: integer
          description: Meme needs to created after this date
        - in: query
          name: name
          schema:
            type: string
          description: Name of meme includes this name
        - in: query
          name: template
          schema:
            type: string
          description: Template of meme equals this template id
        - in: query
          name: tags
          schema:
            type: string
          description: Tags of meme include these tags (JSON string of 1D-string-array)
      responses:
        200:
          description: "successful operation"
          content:
            application/json:
              example:
                msg: "Ok"
                data:
                  - createdAt: "2021-01-14T16:42:53.842Z"
                    views: 3
                    url: "http://localhost:3000/files/download/6000748e"
                    access: "public"
                    mediaType: "image"
                    template: "60007"
                    captions:
                      - text: "my caption"
                        position:
                          top: 0.5
                          left: 0.4
                          right: 0
                          bottom: 0
                        fontFace:
                          fontSize: 40
                          fontFamily: "Roboto"
                          color: "white"
                    id: "34534523"
                    likes:
                      - user: "foo"
                        memeId: "34534523"
                        createdAt: "2021-01-14T16:44:20.285Z"
                        id: "54467567567"
                    comments:
                      - user: "foo"
                        memeId: "34534523"
                        createdAt: "2021-01-14T16:44:20.285Z"
                        id: "54467567567"
                        text: "my comment"
                    tags:
                      - "hello"
                      - "world"
        500:
          description: "Internal server error"
...
  /memes/meme/{id}/:
    get:
      summary: "Get single meme by id"
      operationId: "getSingleMeme"
      security:
        - bearerAuth: []
      tags:
        - Memes
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the meme to get
      responses:
        200:
          description: "Ok"
          content:
            application/json:
              example:
                msg: "Ok"
                data:
                  createdAt: "2021-01-14T16:42:53.842Z"
                  views: 3
                  url: "http://localhost:3000/files/download/6000748e"
                  access: "public"
                  mediaType: "image"
                  template: "60007"
                  captions:
                    - text: "my caption"
                      position:
                        top: 0.5
                        left: 0.4
                        right: 0
                        bottom: 0
                      fontFace:
                        fontSize: 40
                        fontFamily: "Roboto"
                        color: "white"
                  id: "34534523"
                  likes:
                    - user: "foo"
                      memeId: "34534523"
                      createdAt: "2021-01-14T16:44:20.285Z"
                      id: "54467567567"
                  comments:
                    - user: "foo"
                      memeId: "34534523"
                      createdAt: "2021-01-14T16:44:20.285Z"
                      id: "54467567567"
                      text: "my comment"
                  tags:
                    - "hello"
                    - "world"
        400:
          description: "Meme not existing or id not provided"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
...
  /memes/video/:
    get:
      summary: "Get a stream video of random memes"
      operationId: "memeVideo"
      security:
        - bearerAuth: []
      tags:
        - Memes
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: "The number of memes to be included in the video (Default: 100)"
      responses:
        200:
          description: "Ok"
        206:
          description: "Partial Content"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes